+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/lrop_v4_cap_java/app/lrop_v4_cap_java/webapp/test/integration/pages/JourneyRunner.js
@@ -5,7 +5,7 @@ sap.ui.define([
     'use strict';
 
     var runner = new JourneyRunner({
-        launchUrl: sap.ui.require.toUrl('testNameSpace/lropv4capjava') + '/index.html',
+        launchUrl: sap.ui.require.toUrl('testNameSpace/lropv4capjava') + '/test/flpSandbox.html#testNameSpacelropv4capjava-tile',
         pages: {
 			onTheBooksObjectPage: BooksObjectPage
         },