+++ packages/fiori-mcp-server/src/types.ts
@@ -1,4 +1,4 @@
-import type { ApplicationAccess } from '@sap-ux/project-access';
+import type { ApplicationAccess, ProjectType } from '@sap-ux/project-access';
 
 /**
  * Types for Fiori functionality
@@ -52,7 +52,7 @@ export interface GetFunctionalityDetailsInput {
  */
 export interface GetFunctionalityDetailsOutput {
     /** ID of the functionality */
-    id: string;
+    functionalityId: string;
     /** Name of the functionality */
     name: string;
     /** Description of the functionality */
@@ -110,11 +110,13 @@ export interface FioriApp {
     /** Name of the Fiori application */
     name: string;
     /** Path to the Fiori application */
-    path: string;
-    /** Type of the Fiori application */
-    type: 'list-report' | 'freestyle' | 'analytical' | 'overview-page';
-    /** Version of the Fiori application */
-    version: string;
+    appPath: string;
+    /** Path to project */
+    projectPath: string;
+    /** Type of the Fiori project */
+    projectType: ProjectType;
+    /** OData version of the Fiori application */
+    odataVersion: string;
 }
 
 /**
@@ -122,7 +124,7 @@ export interface FioriApp {
  */
 export interface Functionality {
     /** ID or array of IDs for the functionality */
-    id: Array<string | number> | string;
+    functionalityId: Array<string | number> | string;
     /** Description of the functionality */
     description: string;
 }