+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.9.17
+
+### Patch Changes
+
+-   @sap-ux/axios-extension@1.22.7
+
+## 0.9.16
+
+### Patch Changes
+
+-   b62d5b9: fix: make 'bsp' path case insensitive
+
+## 0.9.15
+
+### Patch Changes
+
+-   Updated dependencies [77c1459]
+    -   @sap-ux/store@1.1.4
+
 ## 0.9.14
 
 ### Patch Changes