+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/environment-check
 
+## 0.18.28
+
+### Patch Changes
+
+-   @sap-ux/axios-extension@1.22.7
+
+## 0.18.27
+
+### Patch Changes
+
+-   Updated dependencies [77c1459]
+    -   @sap-ux/store@1.1.4
+    -   @sap-ux/ui5-config@0.29.3
+
 ## 0.18.26
 
 ### Patch Changes