+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.2.10
+
+### Patch Changes
+
+-   Updated dependencies [04d2103]
+    -   @sap-ux/odata-service-inquirer@2.8.0
+    -   @sap-ux/axios-extension@1.22.7
+    -   @sap-ux/inquirer-common@0.7.39
+    -   @sap-ux/abap-deploy-config-inquirer@1.6.13
+
+## 0.2.9
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.6.12
+-   @sap-ux/odata-service-inquirer@2.7.3
+-   @sap-ux/telemetry@0.6.19
+-   @sap-ux/inquirer-common@0.7.38
+
 ## 0.2.8
 
 ### Patch Changes